requirements.txt
beautifulsoup4==4.5.3
beebotte==0.5.0
pkg-resources==0.0.0
PyYAML==3.12
requests==2.12.4
